CP-919 add McSerialPortDevice
