+++ packages/pure/src/device/strategies/pure.strategy.ts
@@ -11,8 +11,8 @@ import {
   RequestConfig,
   Response,
   ResponseStatus,
+  McSerialPortDevice,
 } from "../device.types"
-import { DeviceType } from "../constants"
 import {
   Contact,
   DeviceInfo,
@@ -50,8 +50,8 @@ import { Formatter, FormatterFactory } from "../../formatter"
 export class PureStrategy extends BaseDevice {
   #formatter: Formatter = FormatterFactory.create()
 
-  constructor(path: string, public deviceType: DeviceType) {
-    super(path, deviceType)
+  constructor(baseMcSerialPortDevice: McSerialPortDevice) {
+    super(baseMcSerialPortDevice)
   }
 
   public async connect(): Promise<Response> {
@@ -159,6 +159,5 @@ export class PureStrategy extends BaseDevice {
 }
 
 export const createPureStrategy: CreateDeviceStrategy = (
-  path: string,
-  deviceType: DeviceType
-) => new PureStrategy(path, deviceType)
+  baseMcSerialPortDevice: McSerialPortDevice
+) => new PureStrategy(baseMcSerialPortDevice)
